fix: Exception Error 500 raised for incomplete, invalid format and invalid date of birth  (#291)
* fix: adding new config values to the config values code

* fix: code formatting

* fix: code formatting

* fix: date is now saved as a varchar in the cohort distribution table and in the demograpgpic table to allow for transformation and validation rules to to validation and transformations

* fix: dates is now saved as a varchar in demograpgic but is transformed to a real date time in transform data before getting to cxo cohort dist

* fix: removing test files

* fix: removing test files

* fix: removing white space

+++ application/CohortManager/src/Functions/Shared/Model/Mapping/ParticipantsParquetMap.cs
@@ -16,7 +16,7 @@ public struct ParticipantsParquetMap
     public Int64? NhsNumber;
 
     [MapToColumn("superseded_by_nhs_number")]
-    public Int64? SupersededByNhsNumber;
+    public Int32? SupersededByNhsNumber;
 
     [MapToColumn("primary_care_provider")]
     public string? PrimaryCareProvider;
